requirements.txt
pandas>=1.5
numpy>=1.23
scikit-learn>=1.2
joblib>=1.2
langchain>=0.2
langgraph>=0.2
google-generativeai
xgboost>=1.7
optuna>=3.0
pyyaml>=6.0
autogluon
